+++ pytket/docs/conf.py
@@ -38,9 +38,9 @@ copyright = "2023 Quantinuum"
 author = "Quantinuum"
 
 # The short X.Y version
-version = "1.16"
+version = "1.17"
 # The full version, including alpha/beta/rc tags
-release = "1.16.0"
+release = "1.17.0"
 
 
 # -- General configuration ---------------------------------------------------