+++ tket/include/tket/Circuit/ToffoliBox.hpp
@@ -21,6 +21,17 @@
 namespace tket {
 
 /**
+ * @brief Strategies for synthesising ToffoliBoxes
+ * Matching: use multiplexors to perform parallel swaps on hypercubes
+ * Cycle: use CnX gates to perform transpositions
+ */
+enum class ToffoliBoxSynthStrat { Matching, Cycle };
+
+NLOHMANN_JSON_SERIALIZE_ENUM(
+    ToffoliBoxSynthStrat, {{ToffoliBoxSynthStrat::Matching, "Matching"},
+                           {ToffoliBoxSynthStrat::Cycle, "Cycle"}});
+
+/**
  * @brief Map bitstrings to bitstrings
  *
  */
@@ -35,12 +46,15 @@ class ToffoliBox : public Box {
    * @brief Construct a circuit that synthesise the given state permutation
    *
    * @param permutation map between basis states
+   * @param strat synthesis strategy
    * @param rotation_axis the rotation axis of the multiplexors used in the
-   * decomposition. Can be either Rx or Ry.
+   * decomposition. Can be either Rx or Ry. Only applicable to the
+   * ToffoliBoxSynthStrat::Matching strategy
    *
    */
   explicit ToffoliBox(
       const state_perm_t &permutation,
+      const ToffoliBoxSynthStrat &strat = ToffoliBoxSynthStrat::Matching,
       const OpType &rotation_axis = OpType::Ry);
 
   /**
@@ -75,6 +89,7 @@ class ToffoliBox : public Box {
 
   state_perm_t get_permutation() const;
   OpType get_rotation_axis() const;
+  ToffoliBoxSynthStrat get_strat() const;
 
  protected:
   /**
@@ -85,10 +100,14 @@ class ToffoliBox : public Box {
   void generate_circuit() const override;
 
   ToffoliBox()
-      : Box(OpType::ToffoliBox), permutation_(), rotation_axis_(OpType::Ry) {}
+      : Box(OpType::ToffoliBox),
+        permutation_(),
+        strat_(ToffoliBoxSynthStrat::Matching),
+        rotation_axis_(OpType::Ry) {}
 
  private:
   const state_perm_t permutation_;
+  const ToffoliBoxSynthStrat strat_;
   const OpType rotation_axis_;
 };
 }  // namespace tket